% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bls_test_df.R
\docType{data}
\name{bls_test_df}
\alias{bls_test_df}
\title{BLS API Data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 0 rows and 0 columns.}
\source{
\url{https://www.bls.gov}
}
\usage{
bls_test_df
}
\description{
A dataset from bls API
}
\keyword{datasets}
