% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flights_sqldb_df.R
\docType{data}
\name{flights_sqldb_df}
\alias{flights_sqldb_df}
\title{Flight data from SQL database.}
\format{A data frame with some flight data:
\describe{
  \item{dest}{probably destination}
  \item{delay}{delay maybe?}
  ...
}}
\usage{
flights_sqldb_df
}
\description{
A dataset containing flight data.
}
\keyword{datasets}
